(chore) - POC tree-shakeable preact/utils
